Reimplement GET request
